solar-panel-detection-streamlit-main/requirements.txt
streamlit>=1.28.0
roboflow>=1.1.0
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
# PDF generation dependencies
